Allow setting template display-name
A tiny change so that one has the ability to set a custom name to be displayed in template dropdown.

If statement with is_int() is needed because if you don't set both values the key will always be a number (int), so is_int() works but isset() does not.

Backwards compability:
If no "name", i.e. value is set it will default to "old behaviour" so that it won't break anything.